fix: proper folder and file structure
