+++ app/about/page.tsx
@@ -1,6 +1,6 @@
-import WhoAreWe from "./WhoAreWe";
-import Quest from "./Quest";
-import MissionVisionValues from "./MissionVisionValues";
+import WhoAreWe from "../components/about/WhoAreWe";
+import Quest from "../components/about/Quest";
+import MissionVisionValues from "../components/about/MissionVisionValues";
 export default function About() {
   return (
     <div>
