+++ app/components/about/WhoAreWe.tsx
@@ -3,15 +3,15 @@ import styles from './About.module.css';
 
 export default function WhoAreWe() {
   return (
-    <div className="flex flex-col bg-Tertiary px-4 pt-5 sm:px-5 md:px-10 lg:flex-row ">
+    <div className="flex flex-col bg-Tertiary px-4 sm:px-5 md:px-10 lg:flex-row items-center">
       <div
-        className={`flex-col lg:flex-1 mt-10 leading-none md:text-4xl text-Primary font-AbolitionTestRegular ${styles.lineHeight}`}
+        className={`flex-col lg:flex-1 leading-none text-2xl text-center md:text-3xl lg:text-4xl lg:text-start text-Primary font-AbolitionTestRegular ${styles.lineHeight}`}
       >
         <p className={styles.whoAreWeText}>WHO</p>
         <p className={styles.whoAreWeText}>ARE</p>
         <p className={styles.whoAreWeText}>WE?</p>
       </div>
-      <div className="flex lg:w-3/4 mt-4 pt-5 justify-center items-start lg:pt-1 lg:justify-start lg:px-7">
+      <div className="flex lg:w-3/4 mt-4 pt-10 justify-center items-start lg:pt-1 lg:justify-start lg:px-7">
         <div className="flex flex-col text-center justify-center font-SofiaProMediumItalic text-Primary text-lg md:text-xl">
           <div className="flex flex-col lg:pl-10 pb-5 space-y-4">
             <Image
@@ -19,6 +19,7 @@ export default function WhoAreWe() {
               height={100}
               width={300}
               alt="logo"
+              className=" w-48 md:w-60  lg:w-80"
             />
             <p className="max-w-4xl">
               is a new initiative by the{" "}
